react: make it Celerity friendly
Celerity [1] is a Phabricator subsystem that preprocesses javascripts.
Replace `"development" !== 'production'` to `__DEV__` so Celerity knows how
to optimize it in production mode (when `phabricator.developer-mode` is set
to False), while we can still get developer warnings. Add `@provides react`
so it could be used as a dependence in other scripts.

[1]: https://secure.phabricator.com/book/phabdev/article/celerity/